requirements.txt
transformers
pytorch_pretrained_bert
argparse
tensorflow==2.3.4
argparse
bert-for-tf2
pandas==1.3.2
python-dotenv
google-cloud-aiplatform
python_socketio==5.6.0
aiohttp==3.8.1